Split out Discogs models.
