+++ backend/Discapp.API/Models/Discogs/DiscogsPagination.cs
@@ -6,14 +6,6 @@ namespace Discapp.API.Models.Discogs
 		public int Pages { get; set; }
 		public int PerPage { get; set; }
 		public int Items { get; set; }
-		public required DiscogsUrls Urls { get; set; }
-	}
-
-	public class DiscogsUrls
-	{
-		public string First { get; set; } = string.Empty;
-		public string Last { get; set; } = string.Empty;
-		public string Prev { get; set; } = string.Empty;
-		public string Next { get; set; } = string.Empty;
+		public required DiscogsPaginationUrls Urls { get; set; }
 	}
 }
