+++ backend/Discapp.API/Models/Discogs/DiscogsPaginationUrls.cs
@@ -1,0 +1,10 @@
+namespace Discapp.API.Models.Discogs
+{
+	public class DiscogsPaginationUrls
+	{
+		public string First { get; set; } = string.Empty;
+		public string Last { get; set; } = string.Empty;
+		public string Prev { get; set; } = string.Empty;
+		public string Next { get; set; } = string.Empty;
+	}
+}
