+++ backend/Discapp.API/Models/Discogs/DiscogsWantsReturn.cs
@@ -1,0 +1,8 @@
+namespace Discapp.API.Models.Discogs
+{
+	public class DiscogsWantsReturn
+	{
+		public required DiscogsPagination Pagination { get; set; }
+		public required List<DiscogsReleaseCollection> Wants { get; set; }
+	}
+}
